refactor : MVC 패턴 적용 및 패키지 정리
기존의 코드에서 MVC 패턴을 적용하고 그에 맞도록 패키지를 생성함.
