+++ src/main/java/racingcar/Application.java
@@ -1,6 +1,6 @@
 package racingcar;
 
-import racingcar.domain.GameManager;
+import racingcar.controller.GameManager;
 
 public class Application {
     public static void main(String[] args) {
