+++ src/main/java/racingcar/constant/Error.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.constant;
 
 public enum Error {
     CAR_NAME_LENGTH_ERROR(
