+++ src/test/java/racingcar/model/CarTest.java
@@ -1,9 +1,9 @@
-package racingcar.domain;
+package racingcar.model;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.Test;
-import racingcar.Error;
+import racingcar.constant.Error;
 
 class CarTest {
