+++ src/test/java/racingcar/model/CarsTest.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -8,7 +8,7 @@ import java.util.Arrays;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import racingcar.Error;
+import racingcar.constant.Error;
 
 class CarsTest {
